fix: rename props and import
