+++ src/interfaces/pokemonCard.ts
